+++ nengo_dl/tests/test_simulator.py
@@ -1469,7 +1469,7 @@ def test_synapse_warning(Simulator):
 def test_concat_hang(Simulator, pytestconfig):
     # TODO: check if this is fixed in 1.14
     if ("1.11.0" <= LooseVersion(tf.__version__) < "1.14" and
-            pytestconfig.getoption("--unroll_simulation") > 1):
+            pytestconfig.getoption("--unroll-simulation") > 1):
         pytest.xfail(
             "There is a bug in TensorFlow that causes this test to hang; see "
             "https://github.com/tensorflow/tensorflow/issues/23383")