+++ CHANGES.rst
@@ -25,6 +25,7 @@ Release History
 
 - The default session will now be set to the NengoDL session before calling
   TensorNodes' ``post_build`` function.
+- Renamed the pytest ``unroll_simulation`` argument to ``unroll-simulation``.
 
 **Fixed**